pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pyclvm"
authors = [{name = "BST Labs", email = "bstlabs@caios.io"}]
readme = "README.md"
license = {file = "LICENSE.md"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.8"
dependencies = [
    "boto3 >= 1.18.42",
    "dynacli >= 1.0.8",
    "pyjdict >= 1.0.6",
    "psutil >= 5.9.0",
    "rich >= 12.0.1",
    "sshconf >= 0.2.3",
    "pycryptodome >= 3.14.1",
    "google-cloud-compute >= 1.3.2",
    "ec2map >= 1.0.5",
    "singleton-decorator >= 1.0.0",
    "azure-core >= 1.24.1",
    "azure-common >= 1.1.28",
    "azure-identity >= 1.10.0",
    "azure-mgmt-core >= 1.3.1",
    "azure-mgmt-compute >= 27.1.0",
    "azure-mgmt-subscription >= 3.0.0",
    "oauthlib >= 3.2.0",
]


[project.urls]
Source = "https://github.com/bstlabs/py-clvm"
Home = "https://github.com/bstlabs/py-clvm"


[project.optional-dependencies]
doc = ["mkdocs-material >=8.1.2"]
dev = [
    "black >=22.3.0",
    "pylint >=2.12.2",
    "isort >=5.9.3",
    "autoflake >=1.4",
    "flake8 >=4.0.1",
    "pre-commit >=2.17.0",
    ]
test = [
    "pytest >=5.4.3",
]
gui = [
# Qt5 components could be installed separately, e.g. via brew
    "PyQt5 >= 5.15.6",
    "PyQtWebEngine >= 5.15.5",
]

[project.scripts]
clvm = "pyclvm._clvm:main"

[tool.isort]
profile = "black"
py_version = 39
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json"]
skip_glob = ["docs/*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.flit.sdist]
exclude = ["src/pyclvm/jupyterlab/*.py", "src/pyclvm/web/*.py", "src/pyclvm/rdp.py"]
